+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - strip manageWorkspaces from consumer bit.json by default
 - bug fix - when removing a missing component from a remote, it was showing an obscure error "id.toStringWithoutVersion is not a function"
 - fixed remove local component
+- fix comparing package dependencies
 
 ## [0.12.0-dev.29] - 2018-01-16